requirements.txt
langchain>=1.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph-checkpoint-sqlite>=3.0.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
questionary>=2.0.0
pytest>=7.4.0
pytest-mock>=3.11.1
pytest-cov>=4.1.0

